requirements.txt
Fabric==1.9.0
argparse==1.2.1
backports.ssl-match-hostname==3.4.0.2
ecdsa==0.11
greenlet==0.4.2
motor==0.3
paramiko==1.14.0
pycrypto==2.6.1
pymongo==2.7.1
pytz==2014.4
sleekxmpp==1.3.1
sockjs-tornado==1.0.0
tornado==3.2.2
ujson==1.33
wsgiref==0.1.2
cerberus
prettysettings
hashids==1.1.0
Jinja2
paho-mqtt==1.2